Move config to extension.json and add node groups. (#23)
* Move config to extension.json and add node groups.

Prior to this change, the look of missing nodes and the
physics of the graph were hard-coded and impossible to
change with configuration. This change moves the
configuration to extension.json and adds a "bluelink"
and "redlink" node group. This allows configuration to
be overriden by the options parameter and, ultimately
(once a bug is fixed in MediaWiki core with the behavior
of the array_replace_recursive merge strategy), with the
$wgPageNetworkOptions configuation variable. Note that the
latter will work to add new options but currently not to
modify existing options.

* Fix test.

* Fix tests.

* Fix tests