Update README.md (#54)
* Update README.md

Added an example of how to exclude an additional namespace while preserving the existing ones. I am no good at php so had to look up how to do it.

* Update README.md

Fixed syntax problem, thanks to @cicalese for spotting it. I was so focused on getting the git process right that I didn't pay attention to what I was actually doing :-(

* Update README.md

consistency with previous point